requirements.txt
kokoro==0.7.16
certifi==2025.4.26
fastapi>=0.115.12
flask>=3.1.0
celery>=5.5.1
aiofiles>=24.1.0
uvicorn>=0.34.0
pydantic>=2.10.6
pydantic_core>=2.27.2
setuptools>=75.6.0
sacremoses>=0.0.53
requests>=2.31.0
numpy>=1.24.4
colorama>=0.4.6
python-dotenv>=1.0.0
playsound3>=1.0.0
soundfile>=0.13.1
transformers>=4.46.3
torch>=2.4.1
ollama>=0.4.7
scipy>=1.9.3
soundfile>=0.13.1
protobuf>=3.20.3
termcolor>=2.4.0
pypdf>=5.4.0
ipython>=8.13.0
pyaudio>=0.2.14
librosa>=0.10.2.post1
selenium>=4.27.1
markdownify>=1.1.0
text2emotion>=0.0.5
adaptive-classifier>=0.0.10
langid>=1.1.6
chromedriver-autoinstaller>=0.6.4
httpx>=0.27,<0.29
anyio>=3.5.0,<5
distro>=1.7.0,<2
jiter>=0.4.0,<1
fake_useragent>=2.1.0
selenium_stealth>=1.0.6
undetected-chromedriver>=3.5.5
sentencepiece>=0.2.0
together>=1.5.0
tqdm>4
openai
sniffio
ordered_set
websockets>=12.0
psutil>=5.9.0
# MCP Protocol Dependencies
websockets>=12.0
# MCP Protocol Dependencies
websockets>=12.0
mcp>=1.0.0

# Enhanced Memory and Knowledge Management Dependencies
chromadb>=0.4.0
sentence-transformers>=2.2.0
networkx>=3.0
scikit-learn>=1.3.0
faiss-cpu>=1.7.0
langchain>=0.1.0
tiktoken>=0.5.0
textstat>=0.7.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
pandas>=2.0.0
